+++ app/settings.gradle
@@ -3,4 +3,4 @@ pluginManagement {
 		gradlePluginPortal()
 	}
 }
-rootProject.name = 'modelserver'
+rootProject.name = 'hint'
