+++ app/src/main/kotlin/org/imperial/mrc/hint/ModelserverApplication.kt
@@ -1,4 +1,4 @@
-package org.imperial.mrc.modelserver
+package org.imperial.mrc.hint
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
