+++ app/src/main/kotlin/org/imperial/mrc/hint/controllers/LoginController.kt
@@ -1,4 +1,4 @@
-package org.imperial.mrc.modelserver.controllers
+package org.imperial.mrc.hint.controllers
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Controller
