+++ app/static/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "modelserver",
+  "name": "hint",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
